fixing hash table get

+++ 0x1A-hash_tables/4-hash_table_get.c
@@ -12,7 +12,7 @@
 char *hash_table_get(const hash_table_t *ht, const char *key)
 {
 	hash_node_t *curr;
-	unsigned long int i = 0;
+	int i = 0;
 
 	i = key_index((const unsigned char *)key, ht->size);
 	curr = ht->array[i];
